Updated package name in demo app
