+++ demo/app/src/main/kotlin/de/onecode/compass/demo/details/DetailsDescription.kt
@@ -1,10 +1,10 @@
-package de.onecode.navigator.demo.details
+package de.onecode.compass.demo.details
 
 import de.onecode.compass.api.Destination
 import de.onecode.compass.api.Navigation
 import de.onecode.compass.api.Parameter
 import de.onecode.compass.api.Top
-import de.onecode.navigator.demo.wizard.WizardSubGraph
+import de.onecode.compass.demo.wizard.WizardSubGraph
 
 @Top
 @Destination(name = "Details")
