+++ demo/app/src/main/kotlin/de/onecode/compass/demo/home/SubHomeScreen.kt
@@ -1,4 +1,4 @@
-package de.onecode.navigator.demo.home
+package de.onecode.compass.demo.home
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
