+++ demo/app/src/main/kotlin/de/onecode/compass/demo/theme/Shape.kt
@@ -1,4 +1,4 @@
-package de.onecode.navigator.demo.ui.theme
+package de.onecode.compass.demo.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
