+++ demo/app/src/main/kotlin/de/onecode/compass/demo/theme/Theme.kt
@@ -1,4 +1,4 @@
-package de.onecode.navigator.demo.ui.theme
+package de.onecode.compass.demo.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
