+++ demo/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Navigator Demo</string>
+    <string name="app_name">Compass Demo</string>
 </resources>
